+++ src/Tests/WorkflowValidation.Test/WorkflowBuilderTests.cs
@@ -1,10 +1,4 @@
-﻿using FluentAssertions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿
 namespace WorkflowValidation.Test
 {
     public class WorkflowBuilderTests
@@ -12,69 +6,101 @@ namespace WorkflowValidation.Test
         [Test]
         public void WorkflowBuilder_Basic()
         {
-            WorkflowBuilder.StartWith(() => { }).Should().BeOfType<Workflow>();
+            var builder = new WorkflowBuilder();
+            builder.StartWith(() => { }).Should().BeOfType<Workflow>();
         }
 
         [Test]
         public void WorkflowBuilder_Basic_Then()
         {
-            WorkflowBuilder.StartWith(() => { })
+            var builder = new WorkflowBuilder();
+            builder.StartWith(() => { })
                 .Then(() => { }).Should().BeOfType<Workflow>();
         }
 
         [Test]
         public void WorkflowBuilder_Basic_Step_Type()
         {
-            WorkflowBuilder.StartWith("Step", () => { }).Steps.Single().Should().BeOfType<Step>();
+            var builder = new WorkflowBuilder();
+            builder.StartWith("Step", () => { }).Steps.Single().Should().BeOfType<Step>();
         }
 
         [Test]
         public void WorkflowBuilder_Basic_Step_Title()
         {
-            WorkflowBuilder.StartWith("Step", () => { }).Steps.Single().Name.Should().Be("Step");
+            var builder = new WorkflowBuilder();
+            builder.StartWith("Step", () => { }).Steps.Single().Name.Should().Be("Step");
         }
 
         [Test]
         public void WorkflowBuilder_Context()
         {
-            WorkflowBuilder.StartWith(c => { }).Should().BeOfType<Workflow>();
+            var builder = new WorkflowBuilder();
+            builder.StartWith(c => { }).Should().BeOfType<Workflow>();
         }
 
         [Test]
         public void WorkflowBuilder_Context_Step_Type()
         {
-            WorkflowBuilder.StartWith("Step", c => { }).Steps.Single().Should().BeOfType<Step>();
+            var builder = new WorkflowBuilder();
+            builder.StartWith("Step", c => { }).Steps.Single().Should().BeOfType<Step>();
         }
 
         [Test]
         public void WorkflowBuilder_Context_Step_Title()
         {
-            WorkflowBuilder.StartWith("Step", c => { }).Steps.Single().Name.Should().Be("Step");
+            var builder = new WorkflowBuilder();
+            builder.StartWith("Step", c => { }).Steps.Single().Name.Should().Be("Step");
         }
 
         [Test]
         public void WorkflowBuilder_Then_Context_Type()
         {
-            WorkflowBuilder
-                .StartWith(() => { })
+            var builder = new WorkflowBuilder();
+            builder.StartWith(() => { })
                 .Then(c => { }).Should().BeOfType<Workflow>();
         }
 
         [Test]
         public void WorkflowBuilder_Then_Message_Context_Type()
         {
-            WorkflowBuilder.StartWith(() => { })
+            Workflow.StartWith(() => { })
                 .Then("Message", c => { }).Should().BeOfType<Workflow>();
         }
 
         [Test]
         public void WorkflowBuilder_Count_Steps()
         {
-            WorkflowBuilder.StartWith(() => { })
+            var builder = new WorkflowBuilder();
+            builder.StartWith(() => { })
                 .Then("Message", c => { })
                 .Verify(c => {})
                 
                 .Steps.Should().HaveCount(3);
+        }
+
+        [Test]
+        public void WorkflowBuilder_SetupWorkflow_Reference()
+        {
+            var builder = new WorkflowBuilder();
+            builder.SetupWorkflow(s => { }).Should().BeSameAs(builder);
+        }
+
+        [Test]
+        public void WorkflowBuilder_SetupWorkflow_NotSet()
+        {
+            var builder = new WorkflowBuilder();
+            builder.StartWith(() => { })
+                .WorkflowSetup.Should().BeNull();
+        }
+
+        [Test]
+        public void WorkflowBuilder_SetupWorkflow_Description()
+        {
+            var builder = new WorkflowBuilder();
+            builder.SetupWorkflow(s => s.SetDescription("description"))
+                .StartWith(() => { })
+                .WorkflowSetup.Description.Should().Be("description");
         }
     }
 }