+++ src/WorkflowValidation/Workflow.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Text;
 
 namespace WorkflowValidation
 {
@@ -28,6 +26,7 @@ namespace WorkflowValidation
         {
             _steps = parent.Steps as List<IStep> ?? parent.Steps.ToList();
             Context = parent.Context;
+            WorkflowSetup = parent.WorkflowSetup;
         }
 
         /// <summary>
@@ -39,6 +38,11 @@ namespace WorkflowValidation
         /// Gets the WorkflowContext used in the workflow execution
         /// </summary>
         public WorkflowContext Context { get; set; }
+
+        /// <summary>
+        /// Setup for the workflow
+        /// </summary>
+        public IWorkflowSetup WorkflowSetup { get; set; }
 
         /// <summary>
         /// Add a Step to the workflow
@@ -63,7 +67,12 @@ namespace WorkflowValidation
             }
 
             Context ??= new WorkflowContext();
-            
+
+            if (WorkflowSetup != null)
+            {
+                WorkflowSetup.Run(Context);
+            }
+
             foreach (var step in _steps)
             {
                 Context.CurrentStep = step;
@@ -73,6 +82,69 @@ namespace WorkflowValidation
             }
 
             return this;
+        }
+
+
+
+
+
+
+        /// <summary>
+        /// Start the workflow with the given step
+        /// </summary>
+        /// <param name="step"></param>
+        /// <returns></returns>
+        public static IWorkflowStep StartWith(Action step)
+        {
+            var workflow = new Workflow();
+            workflow.SetStep(new Step(step));
+
+            return workflow;
+        }
+
+        /// <summary>
+        /// Start the workflow with the given step
+        /// </summary>
+        /// <param name="description"></param>
+        /// <param name="work"></param>
+        /// <returns></returns>
+        public static IWorkflowStep StartWith(string description, Action work)
+        {
+            var workflow = new Workflow();
+            workflow.SetStep(new Step(work)
+                .SetName(description)
+            );
+
+            return workflow;
+        }
+
+        /// <summary>
+        /// Start the workflow with the given step
+        /// </summary>
+        /// <param name="step"></param>
+        /// <returns></returns>
+        public static IWorkflowStep StartWith(Action<WorkflowContext> step)
+        {
+            var workflow = new Workflow();
+            workflow.SetStep(new Step(step));
+
+            return workflow;
+        }
+
+        /// <summary>
+        /// Start the workflow with the given step
+        /// </summary>
+        /// <param name="description"></param>
+        /// <param name="work"></param>
+        /// <returns></returns>
+        public static IWorkflowStep StartWith(string description, Action<WorkflowContext> work)
+        {
+            var workflow = new Workflow();
+            workflow.SetStep(new Step(work)
+                .SetName(description)
+            );
+
+            return workflow;
         }
     }
 }