+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## vNext
+- Setup for workflows containing a description that gets output
+  
 ## v1.0.1
 ### Added
 - Add multiple Assertions to Verificationstep